% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{composite_score}
\alias{composite_score}
\title{Compute composite score}
\usage{
composite_score(
  x,
  R,
  mu_x = 100,
  sigma_x = 15,
  mu_composite = 100,
  sigma_composite = 15,
  w = NULL
)
}
\arguments{
\item{x}{Vector of subtest scores}

\item{R}{Subtest score correlation matrix}

\item{mu_x}{Vector of subtest means}

\item{sigma_x}{Vector of subtest standard deviations}

\item{mu_composite}{Composite mean}

\item{sigma_composite}{Composite standard deviation}

\item{w}{Vector of weights}
}
\value{
composite score
}
\description{
Compute composite score
}
\examples{
# Subtest scores
x <- c(12, 14)
R <- matrix(c(1, .6, .6, 1), nrow = 2)
composite_score(
  x = x,
  R = R,
  mu_x = 10,
  sigma_x = 3
)
}
