% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{conditional_covariance}
\alias{conditional_covariance}
\title{Conditional Covariance}
\usage{
conditional_covariance(x, sigma, mu = 0)
}
\arguments{
\item{x}{named numeric vector of predictor scores}

\item{sigma}{named covariance matrix of predictor and outcome variables}

\item{mu}{a single numeric mean for all variables or a named vector of means of predictor and outcome variables}
}
\value{
list of conditional means and a covariance matrix
\itemize{
\item \code{mu_conditional} - The means of the outcome variables conditioned on the values of the predictors in vector x.
\item \code{mu_sigma} - The covariance matrix of the outcome variables conditioned on the values of the predictors in vector x.
\item \code{descriptives_conditional} - A data frame of means and standard deviations of the outcome variables conditioned on the values of the predictors in vector x.
\item \code{x} - The predictor scores from the x parameter
\item \code{sigma} - The unconditional covariance matrix from the sigma parameter
\item \code{mu} - Anamed vector of unconditional means
\item \code{beta} - Regression Coefficients for finding mu_conditional
}
}
\description{
Conditional Covariance
}
\examples{
# Named vector of predictor scores
x <- c(A = 1)

# Named vector of unconditional means
mu <- c(A = 0, B = 0, C = 0)

# Unconditional covariance matrix with row and column names
sigma <- matrix(
  c(
    1, .5, .5,
    .5, 1, .5,
    .5, .5, 1
  ),
  nrow = 3,
  ncol = 3,
  dimnames = list(
    names(mu),
    names(mu)
  )
)

# Conditoinal means and covariance matrix
conditional_covariance(x = x, sigma = sigma, mu = mu)

}
