% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{estimated_stability}
\alias{estimated_stability}
\title{Estimated stability coefficient for IQ}
\usage{
estimated_stability(
  age,
  interval,
  different_battery_family = FALSE,
  b0 = 0.716,
  b1 = 0.003,
  b2 = -0.258,
  b3 = 0,
  b4 = -0.095,
  b5 = -0.138,
  b6 = -0.104
)
}
\arguments{
\item{age}{age at initial testing}

\item{interval}{year between tests}

\item{different_battery_family}{different battery family (e.g., WAIS and Stanford-Binet). Defaults to \code{FALSE}, meaning either the same test or test from the same battery family (e.g., WISC and WAIS))}

\item{b0}{Horizontal asymptote}

\item{b1}{Age scaling factor}

\item{b2}{Age growth rate}

\item{b3}{Interval-age interaction}

\item{b4}{Interval scaling factor}

\item{b5}{Interval growth rate}

\item{b6}{Different battery family effect}
}
\value{
a correlation stability coefficient
}
\description{
Uses best fitting model of general ability from Table 3 from Breit, M., Scherrer, V., Tucker-Drob, E. M., & Preckel, F. (2024). The stability of cognitive abilities: A meta-analytic review of longitudinal studies. Psychological Bulletin. https://doi.org/10.1037/bul0000425
}
\examples{
estimated_stability(
  age = 12,
  interval = 6,
  different_battery_family = TRUE
)
}
