% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{logit2w}
\alias{logit2w}
\title{Convert logits to W scores}
\usage{
logit2w(logit, refw = 500)
}
\arguments{
\item{logit}{numeric vector of logits}

\item{refw}{numeric vector of reference W scores}
}
\value{
numeric vector of W scores
}
\description{
Convert logits to W scores
}
\examples{
logit2w(2)

}
