% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psycheval-package.R
\docType{package}
\name{psycheval-package}
\alias{psycheval}
\alias{psycheval-package}
\title{psycheval: A psychological evaluation toolkit}
\description{
Functions useful in psychological evaluations This package accompanies [*Individual Psychometrics*](https://individual-psychometrics.rbind.io/) online textbook.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjschne/psycheval}
  \item \url{https://wjschne.github.io/psycheval/}
  \item Report bugs at \url{https://github.com/wjschne/psycheval/issues}
}

}
\author{
\strong{Maintainer}: W. Joel Schneider \email{w.joel.schneider@gmail.com} (\href{https://orcid.org/0000-0002-8393-5316}{ORCID})

Authors:
\itemize{
  \item W. Joel Schneider \email{w.joel.schneider@gmail.com} (\href{https://orcid.org/0000-0002-8393-5316}{ORCID})
}

}
\keyword{internal}
