% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rpi}
\alias{rpi}
\title{Convert ability (in W scores by default) to relative proficiency index}
\usage{
rpi(
  x,
  mu = 500,
  scale = 20/log(9),
  criterion = 0.9,
  reverse = FALSE,
  interpretation = FALSE
)
}
\arguments{
\item{x}{numeric vector of ability scores}

\item{mu}{numeric vector of ability scores of reference group}

\item{scale}{number vector of scaling factor. The default value (\code{log(9) / 20}) assumes that x and mu are W scores.}

\item{criterion}{numeric proficiency criterion (between 0 and 1, exclusive)}

\item{reverse}{boolean. If TRUE, the criterion refers to the
proficiency of the person instead of the proficiency of the peer
group. In other words, the role of the x and mu are reversed.}

\item{interpretation}{If TRUE, the rpi's print method will provide an interpretation of
the relative proficiency.}
}
\value{
numeric
}
\description{
Convert ability (in W scores by default) to relative proficiency index
}
\examples{
# What is the probability a person with a W score of 540 can pass
# an item that a person with a 500 W score can pass with a
# probability of .90?
rpi(x = 540, mu = 500, criterion = .9)
# Same as above but with an interpretive statement
rpi(x = 540, mu = 500, criterion = .9, interpretation = TRUE)
# When a person with a W score of 540 has a .9 probability of
# passing an item, what is the probability that a person with a W
# score of 500 will pass it?
rpi(x = 540, mu = 500, criterion = .9, reverse = TRUE, interpretation = TRUE)
}
