% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{standard_score}
\alias{standard_score}
\title{create standard scores}
\usage{
standard_score(
  .data = numeric(0),
  mu = 100L,
  sigma = 15,
  rxx = 0.9,
  ci_level = 0.95,
  accuracy = 1
)
}
\arguments{
\item{.data}{obtained score}

\item{mu}{population mean}

\item{sigma}{population standard deviation}

\item{rxx}{reliability coefficient}

\item{ci_level}{confidence level}

\item{accuracy}{rounding accuracy}
}
\description{
computes confidence intervals, percentiles
}
\section{Slots}{

\describe{
\item{\code{estimated_true_score}}{estimated true score}

\item{\code{standard_error_of_measurement}}{standard error of measurement}

\item{\code{standard_error_of_estimation}}{standard error of estimation}

\item{\code{margin_of_error}}{margin of error}

\item{\code{z}}{z-score associated with \code{ci_level}}

\item{\code{ci_lower_bound}}{lower bound of CI}

\item{\code{ci_upper_bound}}{upper bound of CI}

\item{\code{ci}}{CI as a range}

\item{\code{percentile}}{percentile}
}}

\examples{
x <- standard_score(140, rxx = .97)
x@ci
x@percentile
x@estimated_true_score
}
