% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{w2logit}
\alias{w2logit}
\title{Convert W scores to logits}
\usage{
w2logit(w, refw = 500)
}
\arguments{
\item{w}{numeric vector of W scores}

\item{refw}{numeric vector of reference W scores}
}
\value{
numeric vector of logits
}
\description{
Convert W scores to logits
}
\examples{
w2logit(540)
}
