% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{w2p}
\alias{w2p}
\title{Convert W scores to a probabilities}
\usage{
w2p(w = 500, refw = 500)
}
\arguments{
\item{w}{person ability in w-score units}

\item{refw}{item difficulty in w-score units}
}
\value{
numeric vector of probabilities
}
\description{
Convert W scores to a probabilities
}
\examples{
w2p(w = 520, refw = 500)
}
