% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{w2rpi}
\alias{w2rpi}
\title{Convert W scores to relative proficiency index}
\usage{
w2rpi(w, refw = 500, criterion = 0.9, referse = FALSE, interpretation = FALSE)
}
\arguments{
\item{w}{numeric vector of W scores}

\item{refw}{numeric vector of reference W scores}

\item{criterion}{numeric proficiency criterion}

\item{referse}{boolean. If TRUE, the criterion refers to the proficiency of the person instead of the proficiency of the peer group. In other words, the role of the w and refw are reversed.}

\item{interpretation}{If TRUE, print method will provide an interpretation of
the relative proficiency.}
}
\value{
numeric
}
\description{
Convert W scores to relative proficiency index
}
\examples{
# What is the probability a person with a W score of 540 can pass
# an item that a person with a 500 W score can pass with a
# probability of .90?
w2rpi(w = 540, refw = 500, criterion = .9)
# Same as above but with an interpretive statement
w2rpi(w = 540, refw = 500, criterion = .9, interpretation = TRUE)
# When a person with a W score of 540 has a .9 probability of
# passing an item, what is the probability that a person with a W
# score of 500 will pass it?
w2rpi(w = 540, refw = 500, criterion = .9, referse = TRUE, interpretation = TRUE)
}
