% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{composite_covariance}
\alias{composite_covariance}
\title{Computes covariances of composite scores given a covariance matrix and a weight matrix}
\usage{
composite_covariance(Sigma, w, correlation = FALSE)
}
\arguments{
\item{Sigma}{Covariance matrix}

\item{w}{Weight matrix. Must have the same number of rows as R}

\item{correlation}{If TRUE, return correlations instead of covariances}
}
\description{
Computes covariances of composite scores given a covariance matrix and a weight matrix
}
\examples{
# Create variable names
v_names <- c(paste0("A_", 1:3),paste0("B_", 1:3))
v_composites <- c("A", "B")

# Create covariance matrix
Sigma <- matrix(0.6, nrow = 6, ncol = 6, dimnames = list(v_names, v_names))
diag(Sigma) <- 1

# Create weight matrix
w <- matrix(0, nrow = 6, ncol = 2, dimnames = list(v_names, v_composites))
w[v_names[1:3],"A"] <- 1
w[v_names[4:6],"B"] <- 1
w
# covariance matrix of weighted sums
composite_covariance(Sigma, w)
}
