% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{difference_score}
\alias{difference_score}
\title{Difference score statistics}
\usage{
difference_score(
  x,
  y,
  r_xx = 0.9,
  r_yy = 0.9,
  r_xy = 0,
  mu = 100,
  sigma = 15,
  ci = 0.95,
  mu_x = mu,
  mu_y = mu,
  sigma_x = sigma,
  sigma_y = sigma,
  tails = 2
)
}
\arguments{
\item{x}{first score}

\item{y}{second score}

\item{r_xx}{reliability of x}

\item{r_yy}{reliability of y}

\item{r_xy}{correlation between x and y}

\item{mu}{population mean of both x and y}

\item{sigma}{population standard deviation of both x and y}

\item{ci}{confidence interval of difference score}

\item{mu_x}{population mean of x (defaults to mu)}

\item{mu_y}{population mean of y (defaults to mu)}

\item{sigma_x}{population standard deviation of x (defaults to sigma)}

\item{sigma_y}{population standard deviation of y (defaults to sigma)}

\item{tails}{for significance and prevalance of difference scores}
}
\value{
list
}
\description{
Difference score statistics
}
\examples{
difference_score(
  x = 120,
  y = 110,
  r_xx = .95,
  r_yy = .92,
  r_xy = .65,
  mu = 100,
  sigma = 15)
}
