% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{multivariate_ci}
\alias{multivariate_ci}
\title{General a multivariate confidence interval for a set of scores}
\usage{
multivariate_ci(x, r_xx, mu, sigma, ci = 0.95, v_names = names(x))
}
\arguments{
\item{x}{a vector of scores}

\item{r_xx}{a vector reliability coefficients}

\item{mu}{a vector means}

\item{sigma}{a covariance matrix}

\item{ci}{confidence level}

\item{v_names}{a vector of names}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{variable} - Variable names
\item \code{x} - Variable scores
\item \code{r_xx} - Reliability coefficients
\item \code{mu_univariate} - Expected true score estimated from the corresponding observed score
\item \code{see_univariate} - Standard error of the estimate computed from the corresponding reliability coefficient
\item \code{mu_multivariate} - Expected true score estimated from all observed scores
\item \code{see_multivariate} - Standard error of the estimate computed from the corresponding reliability coefficient
\item \code{upper_univariate} - upper bound of univariate confidence interval
\item \code{lower_univariate} - lower bound of univariate confidence interval
\item \code{upper_multivariate} - upper bound of multivariate confidence interval
\item \code{lower_multivariate} - lower bound of multivariate confidence interval
}
}
\description{
General a multivariate confidence interval for a set of scores
}
\examples{
# Observed Scores
x <- c(
  vci = 130,
  vsi = 130,
  fri = 70,
  wmi = 130,
  psi = 130
)

# Reliability Coefficients
r_xx <- c(
  vci = .92,
  vsi = .92,
  fri = .93,
  wmi = .92,
  psi = .88
)

# Correlation matrix
R <- ("
  index	vci 	vsi 	fri 	wmi 	psi
  vci  	1.00	0.59	0.59	0.53	0.30
  vsi  	0.59	1.00	0.62	0.50	0.36
  fri  	0.59	0.62	1.00	0.53	0.31
  wmi  	0.53	0.50	0.53	1.00	0.36
  psi  	0.30	0.36	0.31	0.36	1.00") |>
  readr::read_tsv() |>
  tibble::column_to_rownames("index") |>
  as.matrix()

# Covariance matrix
sigma <- R * 15^2

# Population means#'
mu <- rep(100, 5)

mci <- multivariate_ci(
  x = x,
  r_xx = r_xx,
  mu = mu,
  sigma = sigma
)

mci

# Conditional covariance of true score estimates
attr(mci, "conditional_covariance")

}
