% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{p_true_score_less_than_threshold}
\alias{p_true_score_less_than_threshold}
\title{Given a true score, what is the probability a true score is below a threshold?}
\usage{
p_true_score_less_than_threshold(x, threshold, rxx, mu = 100, sigma = 15)
}
\arguments{
\item{x}{observed score}

\item{threshold}{threshold score}

\item{rxx}{reliability coefficient (must be between 0 and 1, exclusively)}

\item{mu}{population mean (default = 100)}

\item{sigma}{population standard deviation (default = 15)}
}
\value{
a probability
}
\description{
Assumes multivaraiate normality of true and observed scores
}
\examples{
# What is the probability that a true score is 70 or less when
# the observed score is 65, the reliability coefficient is .95,
# the population mean is 100, and the population standard
# deviation is 15?
p_true_score_less_than_threshold(
   x = 65,
   threshold = 70,
   rxx = .95,
   mu = 100,
   sigma = 15)
}
