% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{x2standard}
\alias{x2standard}
\title{Convert x to a standard score}
\usage{
x2standard(
  x,
  mu_x = mean(x, na.rm = T),
  sigma_x = stats::sd(x, na.rm = T),
  mu_new = 100,
  sigma_new = 15,
  digits = 0,
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{mu_x}{mean of current scores}

\item{sigma_x}{standard deviation of current scores}

\item{mu_new}{mean of new scores}

\item{sigma_new}{standard deviation of new scores}

\item{digits}{rounding digits}
}
\value{
numeric vector
}
\description{
Convert x to a standard score
}
\examples{
x2standard(13, mu_x = 10, sigma_x = 3)

}
